requirements.txt
# Core Deep Learning
torch
torchvision

# Image Processing
Pillow
opencv-python
tifffile
scikit-image

# Data Science & Utilities
numpy
matplotlib
tqdm
scikit-learn

# For Evaluation
torch-fidelity